+++ vitest.config.mjs
@@ -1,0 +1,38 @@
+import { defineConfig } from 'vitest/config'
+import vue from '@vitejs/plugin-vue'
+import { quasar, transformAssetUrls } from '@quasar/vite-plugin'
+import jsconfigPaths from 'vite-jsconfig-paths'
+import { resolve } from 'path'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  test: {
+    environment: 'happy-dom',
+    setupFiles: 'test/vitest/setup-file.js',
+    include: [
+      // Matches vitest tests in any subfolder of 'src' or into 'test/vitest/__tests__'
+      // Matches all files with extension 'js', 'jsx', 'ts' and 'tsx'
+      'src/**/*.vitest.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}',
+      'test/vitest/__tests__/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}',
+    ],
+  },
+  plugins: [
+    vue({
+      template: { transformAssetUrls },
+    }),
+    quasar({
+      sassVariables: 'src/quasar-variables.scss',
+    }),
+    jsconfigPaths(),
+  ],
+  resolve: {
+    alias: {
+      '@': resolve(__dirname, 'src'),
+      stores: resolve(__dirname, 'src/stores'),
+      components: resolve(__dirname, 'src/components'),
+      layouts: resolve(__dirname, 'src/layouts'),
+      pages: resolve(__dirname, 'src/pages'),
+      src: resolve(__dirname, 'src'),
+    },
+  },
+})