+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Added
 
+- Vitest pour les tests unitaires.
 - Initialisation du Rapport de test technique.
 - Changelog pour les changements notables apportés à ce projet.
 
 ### Fixed
 
-- Acces à la page de recherche sans quartier.
+- Accès à la page de recherche sans quartier.